\name{aaf_v1_simple}
\alias{aaf_v1_simple}
\title{Get simple (VTI for domestics) weights from AAF_V1}
\usage{
  aaf_v1_simple(risk)
}
\arguments{
  \item{risk}{A vector of numeric risk levels, from 0 to 1,
  to calculate optimal portfolios from}
}
\value{
  A named vector of asset allocation weights
}
\description{
  Get simple (VTI for domestics) weights from AAF_V1
}
\examples{
aaf_v1_simple(.5)
}
\seealso{
  nothing
}
\keyword{finance,}
\keyword{functions,}
\keyword{little}
\keyword{portfolio}
\keyword{useful}

