% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.alpha.R
\name{add.alpha}
\alias{add.alpha}
\title{Automatically add an alpha transparency level to a color.}
\usage{
add.alpha(col, alpha = 1)
}
\arguments{
\item{col}{input color vector}

\item{alpha}{input transparency level (0 - 100)}
}
\value{
string color vector
}
\description{
Automatically add an alpha transparency level to a color.
}
\examples{
x<-seq(0,100)
y<-sin(x)
cols<-colors(1)[x]
plot(y~x,col=colors(1)[x],pch=19,cex=5)
alpha.cols<-add.alpha(cols,0.5)
plot(y~x,col=alpha.cols,pch=19,cex=5)
}
\seealso{
\code{\link{nchar}} which this function wraps
}
\keyword{alpha}
\keyword{color}
