% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.named.vector.R
\name{as.named.vector}
\alias{as.named.vector}
\title{Make a named vector easily}
\usage{
as.named.vector(vdata, dnames)
}
\arguments{
\item{dnames}{The names to be given to the values.}

\item{ddata}{A vector of data you want to be}

\item{ddata}{The values to be named}
}
\value{
A named vector.
}
\description{
Make a named vector easily
}
\examples{
df <- data.frame(value=seq(1:14), labels=LETTERS[seq(1:14)])
as.named.vector(df$value, df$labels)
}
\author{
Daniel Egan
}

