\name{axis.lim.round}
\alias{axis.lim.round}
\title{Round a number in log space for use in nice axes.}
\usage{
  axis.lim.round(x)
}
\arguments{
  \item{x}{The number to be "nicified" for axes.}
}
\value{
  Another number
}
\description{
  This function is useful for taking a max number from
  data, and converting to a "nice" round number.
}
\examples{
axis.lim.round(pi^seq(1,10,0.2))
}
\seealso{
  Nothing.
}
\keyword{annualize,}
\keyword{convenience}
\keyword{finance,}
\keyword{portfolio,}

