% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/barDiffplot.R
\name{barDiffplot}
\alias{barDiffplot}
\title{A barplot showing differences}
\usage{
barDiffplot(data, cols = c("red", "green", "dark green", space = 0, border =
  NA))
}
\arguments{
\item{data}{A matrix with series in rows, and observations in columns.}

\item{cols}{The plot color for when row 1 is greater than row 2, when row 2 is greater than row 1, and the minimum baseline between them.}
}
\value{
A plot
}
\description{
A matched barplot with values overlaid, and differences higlighted
}
\examples{
data<- data.frame(x=seq(100,10,-10), y=seq(10,100,10))
data<- t(data)
barplot(data)
}
\seealso{
Nothing.
}
\keyword{annualize,}
\keyword{convenience}
\keyword{finance,}
\keyword{portfolio,}

