% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bl.compute.riskaversion.R
\name{bl.compute.riskaversion}
\alias{bl.compute.riskaversion}
\title{Calculates Risk Aversion from a set of benchmark returns and a risk-free rate of return}
\usage{
bl.compute.riskaversion(bench, risk.free = 0)
}
\arguments{
\item{bench}{The vector of benchmark returns}

\item{risk.free}{The numeric or vector of risk free rates of return to remove from the benchmark returns}
}
\value{
risk.aversion The numeric estimate of risk aversion
}
\description{
Calculates Risk Aversion from a set of benchmark returns and a risk-free rate of return
}
\examples{
risky.returns<-rnorm(1000,5,sd=11)
riskfree.returns<-rnorm(1000,1,sd=1)
bl.compute.riskaversion(risky.returns,riskfree.returns)
}
\seealso{
nothing
}
\keyword{economics}
\keyword{finance}
\keyword{forecasting}

