% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deNullArgs.R
\name{deNullArgs}
\alias{deNullArgs}
\title{de-Null the arguments to a parent function.}
\usage{
deNullArgs()
}
\value{
Nothing, modifies parent environment
}
\description{
This function looks at the parent or calling function, and checks if any arguments that have
defaults are currently null. If they are, it replaces the NULL with the default.
}
\examples{
checkFun<- function(x=5){
deNullArgs()
print(x)
}

checkFun(x=NULL)
}
\author{
Daniel Egan
}

