% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/destring.R
\name{destring}
\alias{destring}
\title{Automatically make an object nonstring}
\usage{
destring(x)
}
\arguments{
\item{x}{Input object}
}
\value{
value Non-string object
}
\description{
Automatically add an alpha transparency level to a color.
}
\examples{
destring("24")
}
\seealso{
\code{\link{is.character}} which this function wraps
}
\keyword{alpha}
\keyword{color}

