% Generated by roxygen2 (4.0.2): do not edit by hand
\name{df_names}
\alias{df_names}
\title{Pull dataframe names that match a specific pattern}
\usage{
df_names(pattern, df)
}
\arguments{
\item{pattern}{input string}

\item{df}{input dataframe}
}
\value{
string name vector
}
\description{
Grabs all the names in a data frame which match a specific patter.
}
\examples{
temp<-data.frame(var_1=seq(1:10),var2=seq(1:10),v2=seq(1:10))
df_names("_",temp)
}
\seealso{
\code{\link{nchar}} which this function wraps
}
\keyword{alpha}
\keyword{color}

