% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dollar0.R
\name{dollar0}
\alias{dollar0}
\title{Format numbers as dollars with no cents}
\usage{
dollar0(x)
}
\arguments{
\item{x}{numeric vector to be converted to dollars}
}
\value{
string Vector of formatted strings
}
\description{
Format numbers as dollars with no cents
}
\examples{
dollar0(10^seq(0,4, .5))
dollar0(-10^seq(0,4, .5))
}
\author{
Daniel Egan
}

