% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/elbowArrows.R
\name{elbowArrows}
\alias{elbowArrows}
\title{Arrows with Elbows}
\usage{
elbowArrows(x0, x1, y0, y1, first = "y", ...)
}
\arguments{
\item{x0}{The source x point}

\item{x1}{The target x point}

\item{y0}{The source y point}

\item{y1}{The target y point}

\item{first}{Should the elbow be based on running the y-axis change first, or the x-axis change.}

\item{...}{other graphing parameters.}
}
\value{
A new object with updated names
}
\description{
Puts an arrow with a 90* elbow on a graph
}
\examples{
plot(seq(1,10))
elbowArrows(2,6,2,6, first='y', code=2, col="blue", lwd=4)
elbowArrows(2,6,2,6, first='x', code=2, col="red", lwd=4)
}
\author{
Daniel Egan
}

