% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherVars.R
\name{gatherVars}
\alias{gatherVars}
\title{Gather all variables in the environment.}
\usage{
gatherVars(keep = NULL, exclude = NULL)
}
\arguments{
\item{keep}{Optional. A set of specific vars you want to keep. Cannot specify with exclude.}

\item{exclude}{Optional. A set of specific vars you want to exclude. Cannot specify with keep.}
}
\value{
A list containing all the data.
}
\description{
Gathers all variables in the current environment an makes a list of them. Often used
in conjunction with \code{\link{unpackVars}} to pass arguments back and forth.
}
\examples{
x <- data.frame(x=rep(5,5))
y <- 10

temp<- gatherVars()
temp

tryThis<- function(){
  a<- data.frame(x=c(1,2,3))
  b<- 2
  temp<- gatherVars()
  return(temp)
}
tryThis()

Keep vars
}
\author{
Daniel Egan
}

