% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logMean.R
\name{logMean}
\alias{logMean}
\title{Finds the mean in log-space}
\usage{
logMean(ldata, base = 10)
}
\arguments{
\item{ldata}{A vector of data to take the mean of.}

\item{base}{What base should the logged data be based on? Defaults to 10.}
}
\value{
The log-mean
}
\description{
Takes the data, logs it according to the 'base' input, finds the mean, and raises 'base' to this mean.
}
\examples{
 rdata<- 10^rnorm(30, mean=4,sd=3)
#' Note that means based on skew variables are unreliable. 
format(round(as.numeric(mean(rdata)),0),digits=20,big.mark = ",")
#' LogMeans obey CLT. 
format(round(logMean(rdata),0),big.mark = ",")
}
\author{
Daniel Egan
}
