% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup}
\alias{lookup}
\title{Lookup-like functionality}
\usage{
lookup(term, idata, searchColumns = "all", searchRows = "all",
  outputColumns = "all", outputRows = "all")
}
\arguments{
\item{term}{The search term of the matrix}

\item{idata}{A dataframe or matrix to be searched.}

\item{searchColumns}{Optional - if you only want to find the search term in one specific column, then use this. Accepts names or numbers.}

\item{outputColumns}{Optional - returns all by default, but can be subsetted.}
}
\value{
All matching rows and columns.
}
\description{
Searches a data frame or matrix for a specific term, and returns all matching instances.
}
\examples{
thisData<- data.frame(name=LETTERS[1:10], number=seq(1,10), othername=LETTERS[6:15], stringsAsFactors=FALSE)
thisData
lookup("F", thisData)
lookup("K", thisData, searchColumns="name")
lookup("K", thisData, searchColumns="othername")
lookup("K", thisData, searchColumns="othername", outputColumn="name")
lookup("K", thisData, searchColumns=3, outputColumn=1)
}
\author{
Daniel Egan
}

