% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.mirror.R
\name{matrix.mirror}
\alias{matrix.mirror}
\title{Mirror a matrix,}
\usage{
matrix.mirror(x)
}
\arguments{
\item{x}{The matrix to be rotated}
}
\value{
The rotated matrix
}
\description{
Mirror a matrix,
}
\examples{
x<-matrix(seq(1,9),ncol=3)
matrix.mirror(x)
}
\seealso{
nothing
}
\keyword{asset}
\keyword{classes}
\keyword{colors,}
\keyword{finance,}
\keyword{functions,}
\keyword{little}
\keyword{portfolio,}
\keyword{useful}
