% Generated by roxygen2 (4.0.2): do not edit by hand
\name{matrix.rotate180}
\alias{matrix.rotate180}
\title{Rotate a matrix 180 degrees. See example.}
\usage{
matrix.rotate180(x)
}
\arguments{
\item{x}{The matrix to be rotated}
}
\value{
The rotated matrix
}
\description{
Rotate a matrix 180 degrees. See example.
}
\examples{
x<-matrix(seq(1,9),ncol=3)
matrix.rotate180(x)
}
\seealso{
nothing
}
\keyword{asset}
\keyword{classes}
\keyword{colors,}
\keyword{finance,}
\keyword{functions,}
\keyword{little}
\keyword{portfolio,}
\keyword{useful}

