% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlist.R
\name{nlist}
\alias{nlist}
\title{Named list}
\usage{
nlist(..., labels = TRUE)
}
\arguments{
\item{...}{The variables to be returned}

\item{labels}{TRUE by default.}
}
\value{
List, consisting of the named version of variables.
}
\description{
Returns a named list automatically, rather than losing the names like 'list' does.
}
\examples{
a <- 1:3
b <- 4:6
nlist(a,b)
nlist(a,b,d=0)
}
\references{
llist from HMisc. This is a 100% copy.
}
\author{
Daniel Egan
}
\keyword{capitalgains}
\keyword{tax}
\keyword{tlh}
