% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines.R
\name{plot_lines}
\alias{plot_lines}
\title{Plot all observations as lines against index. The first column is the index.}
\usage{
plot_lines(mat, thisColors = NULL, ...)
}
\arguments{
\item{mat}{matrix or dataframe of data to plot}

\item{thisColors}{A set of colors to use in the plot.}

\item{...}{Further plotting parameters, e.g. col}
}
\value{
A plot
}
\description{
Plot all observations as lines against index. The first column is the index.
}
\examples{
 obs<-20
 rets<-data.frame(a=seq(.7,1.4,length.out=obs),b=runif(obs),c=rnorm(obs))  
 rets$y<-sin(rets$a+rets$b)
 plot_lines(rets)
 rets<-data.frame(matrix(runif(obs*20),ncol=20,nrow=obs))  
 for (i in 1:ncol(rets)) {
 rets[,i]<- rets[,i] + i
 }
 index<-seq(1,obs)
 plot_lines(cbind(index,rets))
}
\seealso{
Nothing.
}
\keyword{annualize,}
\keyword{convenience}
\keyword{finance,}
\keyword{portfolio,}
