% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_html.R
\name{preview_html}
\alias{preview_html}
\title{preview_html}
\usage{
preview_html(x)
}
\arguments{
\item{x}{The object you wish to render in html}
}
\description{
Pushes unrendered output to the RStudio preview pane
}
\examples{
library(knitr)
x <- data.frame(alpha  = LETTERS,
     nums = round(runif(n = length(LETTERS)),2))
kx <- kable(x)
kx
# Doesn't render how you'd actually like.
preview_html(kx)
}
\author{
Daniel Egan
}
