% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replaceName.R
\name{replaceName}
\alias{replaceName}
\title{Replace object name}
\usage{
replaceName(df, oldName, newName)
}
\arguments{
\item{df}{The object with names to be replaced}

\item{oldName}{The old name in the object}

\item{newName}{The replacement name}
}
\value{
A new object with updated names
}
\description{
Replaces the object \code{oldName} with \code{newName}
}
\examples{
thisDF<- data.frame("A" = c(1,2,3), "B"=c(3,4,5))
replaceName(thisDF, "A","C")
}
\author{
Daniel Egan
}
\keyword{BMTFxns}
\keyword{lots}
\keyword{tax}

