% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unpackVars.R
\name{unpackVars}
\alias{unpackVars}
\title{Unpack all vars into the environment.}
\usage{
unpackVars(vars, varlist = NULL)
}
\arguments{
\item{vars}{Optional. A list of variables to be unpacked.}

\item{varlist}{Optional}
}
\value{
Nothing. Changes the environment above.
}
\description{
Gathers all variables in the current environment an makes a list of them. Often used
in conjunction with \code{\link{unpackVars}} to pass arguments back and forth.
}
\examples{
x <- 5
y <- 10
z <- 15

temp<- gatherVars()
adder<- function(stuff){
 unpackVars(stuff, varlist=c('x', 'y'))
 if(!exists('z', inherits=FALSE )) z<- 100
 #' this shows that Z was not unpacked.
 x+y+z
}

adder(temp)
}
\author{
Daniel Egan
}

