% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vlookup.R
\name{vlookup}
\alias{vlookup}
\title{Vlookup-like functionality}
\usage{
vlookup(term, idata, searchColumns = "all", outputColumns = NULL)
}
\arguments{
\item{term}{The search term of the matrix}

\item{idata}{A dataframe or matrix to be searched.}

\item{searchColumns}{Optional - if you only want to find the search term in one specific column, then use this. Accepts names or numbers.}

\item{outputColumns}{Optional - returns all by default, but can be subsetted.}
}
\value{
All matching rows and columns.
}
\description{
Searches a data frame or matrix for a specific term, and returns all matching instances.
}
\examples{
thisData<- data.frame(name=LETTERS[1:10], number=seq(1,10), othername=LETTERS[24:15])
vlookup("U", thisData)
}
\author{
Daniel Egan
}

