% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/XgetDividends.R
\name{XgetDividends}
\alias{XgetDividends}
\title{Get prices & returns from XIgnite API}
\usage{
XgetDividends(symbol, start_date, end_date = as.character(Sys.Date()),
  YAMLfile = NULL, token = NULL, ShowURL = FALSE)
}
\arguments{
\item{symbol}{The symbol to be queries}

\item{start_date}{The beginning of the historical date range}

\item{end_date}{The end of the historical date range. Defaults to today.}

\item{YAMLfile}{A locally held yaml file with XIgnite Credentials.}

\item{token}{Your API Token}

\item{ShowURL}{Should the XIgnite API ULR be shown (helps with debugging)}
}
\value{
An xts object of date and dividends
}
\description{
Get prices & returns from XIgnite API
}
\examples{
# Note: No examples will work without a valid token.
\dontrun{
data<-XgetDividends("AGG","2004-01-01",
   YAMLfile = "~/Documents/Betterment/credentials//XIgniteCredentials.yaml")
head(data,20)
}
}
\author{
Daniel Egan
}
\keyword{annualize}
\keyword{convenience}
\keyword{finance}
\keyword{portfolio}

