% Generated by roxygen2 (4.0.2): do not edit by hand
\name{XgetIndexConstituents}
\alias{XgetIndexConstituents}
\title{Get Index Constituents from XIgnite}
\usage{
XgetIndexConstituents(symbol, YAMLfile = NULL, token = "NA",
  ShowURL = FALSE)
}
\arguments{
\item{symbol}{The index symbol to be use. See \url{http://www.xignite.com/product/XigniteIndexComponents/api/legacy/v1/}}

\item{YAMLfile}{A local YAML formatted file which has your XIgnite credentials}

\item{token}{Your XIgnite API Token, if you want to specify at command line.}

\item{ShowURL}{Should the XIgnite API ULR be shown (helps with debugging)}
}
\value{
A dataframe of date and values
}
\description{
Get Index Constituents from XIgnite
}
\examples{
# Note: no examples will work without a token.
\dontrun{
temp<- XgetIndexConstituents("^DJI",
   YAMLfile = "~/Documents/Betterment/credentials//XIgniteCredentials.yaml")
head(temp)
}
}

