% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.named.vector.R
\name{as.named.vector}
\alias{as.named.vector}
\title{Make a named vector easily}
\usage{
as.named.vector(vdata, dnames)
}
\arguments{
\item{vdata}{The values to be named}

\item{dnames}{The names to be given to the values.}
}
\value{
A named vector.
}
\description{
Make a named vector easily
}
\examples{
df <- data.frame(value=seq(1:14), labels=LETTERS[seq(1:14)])
as.named.vector(df$value, df$labels)
}
\author{
Daniel Egan
}
