% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dollar.R
\name{dollar}
\alias{dollar}
\title{Format numbers as dollars, i.e. adding commas and $ signs}
\usage{
dollar(x, fixed = FALSE, rounder = 2)
}
\arguments{
\item{x}{numeric vector to be converted to dollars}

\item{fixed}{string Should the results be left-aligned?}

\item{rounder}{The digits to be rounded to, before conversion.}
}
\value{
string Vector of formatted strings
}
\description{
Format numbers as dollars, i.e. adding commas and $ signs
}
\examples{
dollar(10^seq(1,10))
dollar(10^seq(1,10),fixed=TRUE)
}
\author{
Daniel Egan
}
\seealso{
\code{\link{nchar}} which this function wraps
}
\keyword{economics}
\keyword{finance}
\keyword{formating}

