% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gatherVars.R
\name{gatherVars}
\alias{gatherVars}
\title{Gather all variables in the environment.}
\usage{
gatherVars(vars = NULL)
}
\arguments{
\item{vars}{Optional. A set of specific vars you want to pass through.}
}
\value{
A list containing all the data.
}
\description{
Gathers all variables in the current environment an makes a list of them. Often used
in conjunction with \code{\link{unpackVars}} to pass arguments back and forth.
}
\examples{
x <- 5
y <- 10

temp<- gatherVars()
temp
}
\author{
Daniel Egan
}

