% Generated by roxygen2 (4.0.2): do not edit by hand
\name{matrix.flip}
\alias{matrix.flip}
\title{Flip a matrix - rotate180, then flip. See example.}
\usage{
matrix.flip(x)
}
\arguments{
\item{x}{The matrix to be rotated}
}
\value{
The rotated matrix
}
\description{
Flip a matrix - rotate180, then flip. See example.
}
\examples{
x<-matrix(seq(1,9),ncol=3)
matrix.flip(x)
}
\seealso{
nothing
}
\keyword{asset}
\keyword{classes}
\keyword{colors,}
\keyword{finance,}
\keyword{functions,}
\keyword{little}
\keyword{portfolio,}
\keyword{useful}

