% Generated by roxygen2 (4.0.2): do not edit by hand
\name{numabbr}
\alias{numabbr}
\title{Abbreviate numbers}
\usage{
numabbr(x)
}
\arguments{
\item{x}{numeric vector to be converted to dollars}
}
\value{
string Vector of formatted strings
}
\description{
Abbreviate numbers
}
\examples{
options(scipen=10)
x<- 10^seq(1,10)*1.3
unlist(sapply(x, numabbr))
}
\seealso{
\code{\link{nchar}} which this function wraps
}
\keyword{economics}
\keyword{finance}
\keyword{formating}

