\name{removeWhitespace}
\alias{removeWhitespace}
\title{Remove whitespace from strings.}
\usage{
removeWhitespace(x)
}
\arguments{
  \item{x}{input string with whitespace to be removed.}
}
\value{
\code{string} A string with now whitespace...
}
\description{
Remove whitespace from strings.
}
\examples{
removeWhitespace("   blah    ")
}
\seealso{
\code{\link{gsub}}
}
\keyword{string}
\keyword{trim}

