% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeWhitespace.R
\name{removeWhitespace}
\alias{removeWhitespace}
\title{Remove whitespace from strings.}
\usage{
removeWhitespace(x)
}
\arguments{
\item{x}{input string with whitespace to be removed.}
}
\value{
\code{string} A string with now whitespace...
}
\examples{
removeWhitespace("   blah    ")
}
\seealso{
\code{\link{gsub}}
}
\keyword{string}
\keyword{trim}

