% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Round to two places. Just a convenience thang...}
\usage{
round2(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A vector of rounder values.
}
\description{
Round to two places. Just a convenience thang...
}
\examples{
round2(runif(10))
}
\seealso{
nothing
}
\keyword{functions,}
\keyword{little}
\keyword{percent,}
\keyword{useful}

