% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unpackEllipse.R
\name{unpackEllipse}
\alias{unpackEllipse}
\title{Unpack arguments from ellipse}
\usage{
unpackEllipse(params)
}
\arguments{
\item{params}{An ellipse style parameters}
}
\value{
contents of ellipses as objects to higher environment.
}
\description{
A convenience function for unpacking the contents of ellipses in a function input. Writes the ellipses to the
higher environment.
}
\examples{
print("hello")
}

