% Generated by roxygen2 (4.0.2): do not edit by hand
\name{XgetSymbols}
\alias{XgetSymbols}
\title{Get prices & returns from XIgnite API}
\usage{
XgetSymbols(symbol, start_date, end_date = as.character(Sys.Date()), adjtype,
  quotetype = "LastClose", YAMLfile = NULL, token = NULL,
  ShowURL = FALSE)
}
\arguments{
\item{symbol}{The symbol to be queries}

\item{start_date}{The beginning of the historical date range}

\item{end_date}{The end of the historical date range. Defaults to today.}

\item{adjtype}{The type of adjustment, if any, to be made. Potentially
\itemize{
  \item{None}
  \item{SplitOnly}
  \item{CashDividendOnly}
  \item{SplitAndProportionalCashDividend}
  \item{SplitAndProportionalCashDividend}
  \item{SplitAndCashDividend}
  \item{All}
}}

\item{quotetype}{The type of data to be returns. Options are:
\itemize{
  \item{Last}
  \item{Open}
  \item{High}
  \item{Low}
  \item{Volume}
  \item{Close (the default)}
  \item{PercentChangeFromOpen}
  \item{ChangeFromLastClose}
  \item{PercentChangeFromLastClose}
}}

\item{YAMLfile}{A locally held yaml file with XIgnite Credentials.}

\item{token}{Your XIgnite API Token}

\item{ShowURL}{Should the XIgnite API ULR be shown (helps with debugging)}
}
\value{
An xts object of date and values
}
\description{
Get prices & returns from XIgnite API
}
\examples{
# Note: Examples will no work without a valid XIgnite token.
\dontrun{
data<-XgetSymbols("EEM", start_date="2013-01-01", adjtype="SplitOnly",
   YAMLfile = "~/Documents/Betterment/credentials/XIgniteCredentials.yaml")
head(data,20)
}
}
\author{
Daniel Egan
}
\seealso{
Nothing.
}
\keyword{annualize,}
\keyword{convenience}
\keyword{finance,}
\keyword{portfolio,}

